# src/main/resources/api.yml
openapi: 3.0.0
info:
  version: 1.0.0
  title: AccountService
servers:
  - url: https://localhost:8080/api
paths:
  /accounts:
    get:
      summary: Get list of accounts
      responses:
        "200":
          description: Successful request
          content:
            application/json:
              schema:
                $ref: './schemas/accounts.schema.json'
        "404":
          description: Nothing found for your request
          content:
            application/json:
              schema:
                $ref: './schemas/errorResource.schema.json'
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: './schemas/errorResource.schema.json'

    post:
      summary: Create new account
      requestBody:
        content:
          application/json:
            schema:
              $ref: './schemas/account.schema.json'
      responses:
        "201":
          description: Object created successfully
        "400":
          description: Bad request. The provided input does not comply to the contract.
          content:
            application/json:
              schema:
                $ref: './schemas/errorResource.schema.json'
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: './schemas/errorResource.schema.json'
  /accounts/{accountId}:
    description: Get account details
    get:
      description: Retrieve account details
      parameters:
        - name: accountId
          in: path
          required: true
          description: The id of the account
          schema:
            type: string
      responses:
        "200":
          description: Successful request
          content:
            application/json:
              schema:
                $ref: './schemas/account.schema.json'
        "404":
          description: Nothing found for your request
          content:
            application/json:
              schema:
                $ref: './schemas/errorResource.schema.json'
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: './schemas/errorResource.schema.json'
    delete:
      description: Delete account
      parameters:
        - name: accountId
          in: path
          required: true
          description: The id of the account
          schema:
            type: string
      responses:
        "204":
          description: Account deleted
        "404":
          description: Nothing found for your request
          content:
            application/json:
              schema:
                $ref: './schemas/errorResource.schema.json'
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: './schemas/errorResource.schema.json'